pyproject.toml
[tool.poetry]
name = "mct-search"
version = "0.1.0"
description = ""
authors = ["김성동 <sdkim@didim365.co.kr>"]
readme = "README.md"
packages = [
    { include = "app" },
]


[tool.poetry.dependencies]
python = "^3.10"
openai = "^1.61.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
